backend/requirements.txt
flask
flask-cors
pandas
numpy
requests
groq
python-dotenv
